Update full_pipeline.py
The cache directory is cleared to prevent it from filling up. A more sophisticated version would remove just files added during the loop, but I opted for removing all the images. So that could be improved. Also, there is new logic for checking if expression files have been downloaded in the case of expression flag is true.